requirements.txt
pymupdf
openai-whisper
langchain
langchain-community
langchain-experimental
langchain-huggingface
langchain-chroma
langchain-ollama
pydub
